+++ portal/routes.py
@@ -33,13 +33,19 @@ def details(type, uuid):
     if type not in types:
         abort(404)
     client = ApiClient('TODO: base url from config')
-    details = client.get_entity(uuid)
-    details_html = object_as_html(details)
+
+    details_html = object_as_html(client.get_entity(uuid))
+    provenance = client.get_provenance(uuid)
+
     if type in {'file'}:  # TODO: As we have other specializations, add them here.
         template = f'pages/details/details_{type}.html'
     else:
         template = f'pages/details/details_base.html'
-    return render_template(template, types=types, type=type, uuid=uuid, details_html=details_html)
+    return render_template(
+        template, types=types, type=type, uuid=uuid,
+        details_html=details_html,
+        provenance=provenance
+    )
 
 
 @blueprint.route('/help')